// web/src/shared/api/generated/model/getUsersParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export type GetUsersParams = {
search?: string;
status?: UserStatus;
role?: UserRole;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageNumber?: number | string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageSize?: number | string;
};